quic: move Quic state into quic module
