+++ rustls/src/server/mod.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "quic")]
-use crate::conn::Quic;
 use crate::conn::{Connection, ConnectionCommon, IoState, PlaintextSink, Reader, Writer};
 use crate::error::Error;
 use crate::key;
@@ -11,7 +9,7 @@ use crate::msgs::enums::ProtocolVersion;
 use crate::msgs::enums::SignatureScheme;
 use crate::msgs::handshake::ServerExtension;
 #[cfg(feature = "quic")]
-use crate::quic;
+use crate::quic::{self, Quic};
 use crate::sign;
 use crate::suites::SupportedCipherSuite;
 use crate::verify;
